Check for the array-ness of $value before trying to use it as such to avoid an E_WARNING in certain circumstances
Namely with the NewsBox extension installed, an incomplete navigation menu item (with "newsbox-title" being its title -- i.e. that string literally instead of being treated as a msg key) would be output and its contents would be this:
PHP Warning:  Invalid argument supplied for foreach() in /var/www/html/135/skins/Poncho/Poncho.phtml on line 50

With this patch, NewsBox as well as any and all other extensions which try to output raw HTML in the navigation area are silently ignored and no E_WARNING gets emitted.

Change-Id: I2fa7617fc72f772b5c0b20003e4f82fb3f605659